Discord epoch and message state order support
Added support for Discord epochs instead of embed timestamps. The reasoning for this is to show logs down to the second. Discord embed timestamps don't seem to offer it. Placeholder messages are now generated (configurable string in dotenv) and then edited with the embed data after it's generated. This allows for exact times for message generation. Also added error correction that gets logged in the message whenever there is null values. I'm not sure if this is a bug, limitation, or an API issue, but I can't seem to find anything online related to this problem.