null channel ID bug workaround and guild matching
This is a work around to get around when a lot of events happen all at once. Caching the channel ID into a different variable enforces that the channel data is always available. Leaving previous workaround code in as a backup in case this becomes an issue in the future. Also added single server enforcement.