Query audit logs on init
Now audit logs are queried on load and if audit logs are found within the (ridiculous) 5 minute threshold, it will add them to the global lookup so that external move events can still be accurately logged.

I found a bug in where when a user is moved back and forth in a bunch of different scenarios to the same channel, sometimes it will be flagged incorrectly by checkForExternalMove. It only happens in very specific scenarios where I'm surprised it works at all, so it is good enough for now unless it comes back to haunt me.